+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## 3.22.0 (2022-06-08)
 
 ### Adds